pyproject.toml
[tool.poetry]
name = "mam"
version = "0.1.0"
description = "API RESTful de Gerenciamento de Consultas Médicas"
authors = ["jaqueline <jaquelinesa.82@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"
faker = "^37.1.0"
psycopg2-binary = "^2.9.10"
dj-database-url = "^2.3.0"
python-dotenv = "^1.1.0"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.1.1"
pytest = "^8.3.5"
pytest-django = "^4.11.1"
coverage = "^7.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
